+++ src/main/java/blackjack/model/card/Deck.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.card;
 
 import java.util.ArrayDeque;
 import java.util.ArrayList;
