+++ src/main/java/blackjack/model/card/Hand.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.card;
 
 import java.util.ArrayList;
 import java.util.List;
