+++ src/main/java/blackjack/model/card/Rank.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.card;
 
 public enum Rank {
     ACE(1),
