+++ src/main/java/blackjack/model/game/Answer.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.game;
 
 public enum Answer {
     YES,
