+++ src/main/java/blackjack/model/game/BlackjackResult.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.game;
 
 public enum BlackjackResult {
     WIN,
