+++ src/main/java/blackjack/model/participant/Players.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.participant;
 
 import java.util.Arrays;
 import java.util.Iterator;
