+++ src/main/java/blackjack/view/dto/ResultDto.java
@@ -1,6 +1,6 @@
 package blackjack.view.dto;
 
-import blackjack.model.BlackjackResult;
+import blackjack.model.game.BlackjackResult;
 
 public record ResultDto(
         String playerName,
