+++ src/test/java/blackjack/model/card/DeckTest.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
