+++ src/test/java/blackjack/model/participant/DealerTest.java
@@ -1,7 +1,10 @@
-package blackjack.model;
+package blackjack.model.participant;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
+import blackjack.model.card.Card;
+import blackjack.model.card.Rank;
+import blackjack.model.card.Suit;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
